workflow: add support for ssh_pass_file
Ideally all targets would be using SSH key-based authentication.

In practice (especially when working with embedded device) this is not
always feasible without having to re-flash the target.

The 'ssh_password' option can already be used in such a
case. However, it requires to store the password in a git repository,
which is often not desirable (even for passwords that are not secret).

To slightly improve it, add 'ssh_pass_file'. If provided, it will
pass the '-f' option to sshpass, which will make it read the password
from a file (that could be located outside any git repository).

Note that the 'ssh_pass_file' option is not compatible with the
'ssh_password' option, but by simplicity no check is done for that.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>
(cherry picked from commit 33c594e5324a340e838fea4164d174e7139f9873)