workflow: support for ssh_password and ssh_pass_file for remote scripts
So far, execute_shell_script_remote was not making use of ssh_password
and ssh_pass_file despite the fact that they are exposed in the
options.

This is problematic in cases where ssh requires a password, as scp
would need it as well.

Fix it by adding an sshpass prefix similar to how it is done for ssh,
when either ssh_password or ssh_pass_file is provided.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>